refactor(native): split out upstream std.rs
